requirements.txt
# General
python-decouple
pydantic-settings
pydantic
Jinja2

# LLM / AI
openai
groq